Add performance tools EXecute NOP
This tool allocate a memory region with X attribute, then fill it
with NOP instructions, and fill a RET instruction at the end of
this region. After filling, this program jump to this region and
execute NOPs in it.
This tool is used to benchmark cpu execution speed across numa
nodes.

You can use K, M and G for a human readable input.
eg.
exnop 1G # means allocate an 1G byte region

You may need execute this program with numactl for better control.